+++ src/Middleware/CharacterSolver.php
@@ -7,13 +7,14 @@ use Illuminate\Foundation\Application;
 
 /**
  * CharacterSolver Middleware
+ *
  * @link https://gist.github.com/blueskan/2d82af0204b4fc4b0f14
- * @package App\Http\Middleware
+ * @package Juy\CharacterSolver
  */
 class CharacterSolver {
 
     /**
-     * Create a new middleware instance.
+     * Create a new middleware instance
      *
      * @param Application $app
      */
@@ -23,7 +24,7 @@ class CharacterSolver {
     }
 
     /**
-     * Handle an incoming request.
+     * Handle an incoming request
      *
      * @param  \Illuminate\Http\Request  $request
      * @param  \Closure  $next
@@ -32,7 +33,9 @@ class CharacterSolver {
     public function handle($request, Closure $next)
     {
         $response = $next($request);
-        $response->setContent(strtr($response->getContent(), $this->app['config']->get('charactersolver.translate')));
+        $response->setContent(
+            strtr($response->getContent(), $this->app['config']->get('charactersolver.translate'))
+        );
 
         return $response;
     }