+++ src/ServiceProvider.php
@@ -2,25 +2,42 @@
 
 namespace Juy\CharacterSolver;
 
-use Illuminate\Support\ServiceProvider as IlluminateServiceProvider;
+use Illuminate\Support\ServiceProvider as BaseServiceProvider;
 
-class ServiceProvider extends IlluminateServiceProvider
+/**
+ * Class ServiceProvider
+ * 
+ * @package Juy\CharacterSolver
+ */
+class ServiceProvider extends BaseServiceProvider
 {
     /**
-     * Register the application services.
+     * Package name
+     *
+     * @const string
+     */
+    const PACKAGE_NAME = 'charactersolver';
+
+    /**
+     * Indicates if loading of the provider is deferred
+     *
+     * @var bool
+     */
+    protected $defer = false;
+
+    /**
+     * Register the application services
      *
      * @return void
      */
     public function register()
     {
         // Default package configuration
-        $this->mergeConfigFrom(
-            __DIR__.'/config/charactersolver.php', 'charactersolver'
-        );
+        $this->mergeConfig();
     }
 
     /**
-     * Perform post-registration booting of services.
+     * Perform post-registration booting of services
      *
      * @return void
      */
@@ -41,12 +58,12 @@ class ServiceProvider extends IlluminateServiceProvider
             return;
         }
 
-        // Add a new middleware to end of the stack if it does not already exist.
+        // Add a new middleware to end of the stack if it does not already exist
         $this->registerMiddleware(Middleware\CharacterSolver::class);
     }
 
     /**
-     * Register the CharacterSolver middleware.
+     * Register the CharacterSolver middleware
      *
      * @param  string $middleware
      */
@@ -57,12 +74,38 @@ class ServiceProvider extends IlluminateServiceProvider
     }
 
     /**
-     * Publish the config file.
+     * Default package configuration
+     *
+     * @return void
+     */
+    protected function mergeConfig()
+    {
+        $this->mergeConfigFrom(
+            $this->packagePath('config/config.php'), self::PACKAGE_NAME
+        );
+    }
+
+    /**
+     * Publish the config file
+     *
+     * @return void
      */
     protected function publishConfig()
     {
         $this->publishes([
-            __DIR__.'/config/charactersolver.php' => config_path('charactersolver.php')
+            $this->packagePath('config/config.php') => config_path(self::PACKAGE_NAME . '.php')
         ], 'config');
     }
+    
+    /**
+     * Loads a path relative to the package base directory
+     *
+     * @param string $path
+     * @return string
+     */
+    protected function packagePath($path = '')
+    {
+        return sprintf('%s/../%s', __DIR__, $path);
+    }
+
 }